Repository Summary
This repository contains the bulk of the projects I had worked on during the duration of the HTML and CSS course. This repository contains assignments of varying depth depending on what we were working on at the time. It contains varying degrees of HTML and CSS development. Progress in this course progressed from simple HTML, to HTML with internal CSS and then we learned about how to do external CSS links to simplify the presentation of the HTML syntax. We had also learned how to build a page utilizing bootstrap which showed us how to simplify building a page.